Selenium_migration/requirements.txt
aiogram==3.4.1
APScheduler==3.10.4
psycopg2-binary==2.9.9
selenium==4.15.2
python-dotenv==1.0.1
openai==2.8.1
redis==7.1.0